domain-expiry/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
python-dateutil==2.9.0.post0
python-whois==0.9.4
